+++ src/components/StyleExampleModal/StyleExampleModal.module.scss
