+++ src/pages/RecognitionFailPage/RecognitionFailPage.module.scss
