+++ src/pages/SelectStylePage/SelectStylePage.module.scss
